# api-contract.openapiv3.yaml
openapi: 3.0.3
info:
  title: Regents Mobile API
  version: 0.2.0
  description: |
    Source-of-truth contract for the current Regents Mobile backend surface.

    Mobile wallet flows are live. Mobile Regent records are sourced from Platform projection.
    Talk records are sourced from Platform Regent Work Runtime. The mobile backend only owns
    wallet intent and receipt state needed for phone wallet rails.
...
paths:
...
  /mobile/agents/{agent_id}/voice/status:
    get:
      operationId: getMobileHermesVoiceStatus
      summary: Check whether Hermes voice is ready for one mobile Regent
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agent_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hermes voice status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HermesVoiceStatus'
        '403':
          description: ChatGPT connection is required before voice starts
          x-regent-error-codes: [ChatGptAccountRequired, HermesVoiceUnavailable]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /mobile/agents/{agent_id}/voice/session:
    post:
      operationId: createMobileHermesVoiceSession
      summary: Create one short-lived Hermes voice session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agent_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHermesVoiceSessionRequest'
      responses:
        '201':
          description: Short-lived Hermes voice session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HermesVoiceSession'
        '403':
          description: ChatGPT connection is required before voice starts
          x-regent-error-codes: [ChatGptAccountRequired, HermesVoiceUnavailable]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /mobile/agents/{agent_id}/voice/prewarm:
    post:
      operationId: prewarmMobileHermesVoice
      summary: Prepare Hermes voice for one mobile Regent
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agent_id
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Hermes voice prewarm accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        '403':
          description: ChatGPT connection is required before voice starts
          x-regent-error-codes: [ChatGptAccountRequired, HermesVoiceUnavailable]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /mobile/agents/{agent_id}/voice/disconnect:
    post:
      operationId: disconnectMobileHermesVoice
      summary: End one Hermes voice session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agent_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              additionalProperties: false
              properties:
                session_id:
                  type: string
      responses:
        '200':
          description: Hermes voice session ended
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        '404':
          description: Voice session was not found
          x-regent-error-codes: [VoiceSessionNotFound]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /mobile/agents/{agent_id}/voice/tool-results:
    post:
      operationId: submitMobileHermesVoiceToolResult
      summary: Submit one iOS voice tool result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agent_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HermesVoiceToolResult'
      responses:
        '202':
          description: Hermes voice tool result accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        '403':
          description: ChatGPT connection is required before voice results are accepted
          x-regent-error-codes: [ChatGptAccountRequired, HermesVoiceUnavailable]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Voice session was not found
          x-regent-error-codes: [VoiceSessionNotFound]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    RateLimitError:
      description: Too many requests
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    HermesVoiceAccount:
      type: object
      additionalProperties: false
      required: [required, satisfied, provider]
      properties:
        required:
          type: boolean
          enum: [true]
        satisfied:
          type: boolean
        provider:
          type: string
          enum: [openai_chatgpt]
        connect_url:
          type: string
          format: uri
          nullable: true
...
    CreateHermesVoiceSessionRequest:
      type: object
      additionalProperties: false
      required: [device_capabilities]
      properties:
        voice:
          type: string
          default: marin
        locale:
          type: string
        timezone:
          type: string
        reasoning_effort:
          type: string
          enum: [low, medium, high]
          default: low
        device_capabilities:
          type: array
          items:
            type: string
        preferred_transport:
          type: string
          enum: [webrtc]
          default: webrtc
    HermesVoiceSession:
      type: object
      additionalProperties: false
      required: [session_id, agent_id, realtime, tools, expires_at, account]
      properties:
        session_id:
          type: string
        agent_id:
          type: string
        expires_at:
          type: string
          format: date-time
        account:
          $ref: '#/components/schemas/HermesVoiceAccount'
        realtime:
          type: object
          additionalProperties: false
          required: [provider, model, client_secret, client_secret_expires_at]
          properties:
            provider:
              type: string
              enum: [openai-realtime]
            model:
              type: string
              default: gpt-realtime-2
            client_secret:
              type: string
            client_secret_expires_at:
              type: string
              format: date-time
            calls_url:
              type: string
              default: https://api.openai.com/v1/realtime/calls
        tools:
          type: array
          items:
            $ref: '#/components/schemas/HermesVoiceTool'
    HermesVoiceStatus:
      type: object
      additionalProperties: false
      required: [enabled, health, agent_id, account]
      properties:
        enabled:
          type: boolean
        health:
          type: string
          enum: [ok, degraded, unavailable]
        agent_id:
          type: string
        account:
          $ref: '#/components/schemas/HermesVoiceAccount'
        active_session_id:
          type: string
          nullable: true
        active_turn_id:
          type: string
          nullable: true
        queue_depth:
          type: integer
        last_event_id:
          type: string
          nullable: true
    HermesVoiceTool:
      type: object
      additionalProperties: false
      required: [name, owner, description]
      properties:
        name:
          type: string
        owner:
          type: string
          enum: [hermes, ios]
        description:
          type: string
        requires_approval:
          type: boolean
          default: false
    HermesVoiceToolResult:
      type: object
      additionalProperties: false
      required: [session_id, tool_call_id, ok]
      properties:
        session_id:
          type: string
        tool_call_id:
          type: string
        ok:
          type: boolean
        result: {}
        error:
          type: object
          nullable: true
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, product, status, path, request_id, message]
          properties:
            code:
              type: string
            product:
              type: string
              enum: [ios]
            status:
              type: integer
            path:
              type: string
            request_id:
              type: string
              nullable: true
            message:
              type: string
            next_steps:
              type: string
              nullable: true
          additionalProperties: false